Analytics: move code to analyticsProcessor
